+++ src/ts/src/public/project/ProjectAddAdministratorRequest.ts
@@ -1,0 +1,9 @@
+import { ProjectItemDto } from "./dto";
+
+export interface ProjectAddAdministratorRequest {
+  id: string;
+  accountId: string;
+}
+export interface ProjectAddAdministratorResponse {
+  item: ProjectItemDto;
+}