+++ src/ts/src/public/project/ProjectCreateRequest.ts
@@ -1,0 +1,8 @@
+import { ProjectItemDto } from "./dto";
+
+export interface ProjectCreateRequest {
+  name: string;
+}
+export interface ProjectCreateResponse {
+  item: ProjectItemDto;
+}