+++ src/ts/src/public/project/ProjectDetachResourceRequest.ts
@@ -1,0 +1,9 @@
+import { ProjectItemDto } from "./dto";
+
+export interface ProjectDetachResourceRequest {
+  id: string;
+  resourceId: string;
+}
+export interface ProjectDetachResourceResponse {
+  item: ProjectItemDto;
+}