+++ src/ts/src/public/project/ProjectGetRequest.ts
@@ -1,0 +1,5 @@
+import { ProjectItemDto } from "./dto";
+
+export interface ProjectGetResponse {
+  item: ProjectItemDto;
+}