+++ src/ts/src/public/project/ProjectPotentialResourceListRequest.ts
@@ -1,0 +1,5 @@
+import { ProjectPotentialResourceItemDto } from "./dto";
+
+export interface ProjectPotentialResourceListResponse {
+  items: ProjectPotentialResourceItemDto[];
+}