+++ src/ts/src/public/project/ProjectRemoveAdministratorRequest.ts
@@ -1,0 +1,9 @@
+import { ProjectItemDto } from "./dto";
+
+export interface ProjectRemoveAdministratorRequest {
+  id: string;
+  administratorId: string;
+}
+export interface ProjectRemoveAdministratorResponse {
+  item: ProjectItemDto;
+}