+++ src/ts/src/public/project/ProjectUpdateRequest.ts
@@ -1,0 +1,9 @@
+import { ProjectItemDto, ProjectUpdatedValueDto } from "./dto";
+
+export interface ProjectUpdateRequest {
+  id: string;
+  values: ProjectUpdatedValueDto;
+}
+export interface ProjectUpdateResponse {
+  item: ProjectItemDto;
+}