+++ src/ts/src/public/project/index.ts
@@ -1,0 +1,15 @@
+export * from "./api";
+export * from "./dto";
+export * from "./values";
+
+// All the requests
+export * from "./ProjectAddAdministratorRequest";
+export * from "./ProjectAttachResourceRequest";
+export * from "./ProjectCreateRequest";
+export * from "./ProjectDeleteRequest";
+export * from "./ProjectDetachResourceRequest";
+export * from "./ProjectGetRequest";
+export * from "./ProjectListRequest";
+export * from "./ProjectPotentialResourceListRequest";
+export * from "./ProjectRemoveAdministratorRequest";
+export * from "./ProjectUpdateRequest";