+++ src/ts/src/public/project/values.ts
@@ -1,0 +1,6 @@
+/**
+ * A limit on why a potential resource can't be used in a project.
+ * - "RequireUpgrade": The resource must be upgraded before it can be used in a
+ *   project
+ */
+export type ProjectPotentialResourceLimit = "RequireUpgrade";