+++ src/ts/src/public/workspace/dto.ts
@@ -1,19 +1,28 @@
-import {
-  WorkspaceMemberRole,
-  WorkspaceType,
-  WorkspaceStatus,
-} from "./values";
+import { WorkspaceMemberRole, WorkspaceType, WorkspaceStatus } from "./values";
 import {
   TaskPropertyDefinitionDto,
   ValueDisplayType,
 } from "../task-property-definition";
 import { MapDisplaySettings } from "../map";
 
+export interface WorkspaceMemberAuthorizationItem {
+  /**
+   * If set to `true` then this member is directly part of this project.
+   */
+  direct: boolean;
+
+  /**
+   * List of projectIds that give this member indirect access to this project.
+   */
+  byProjectIds: string[];
+}
+
 export interface WorkspaceMemberItem {
   accountId: string;
   displayName: string;
   role: WorkspaceMemberRole;
   since: Date;
+  authorization: WorkspaceMemberAuthorizationItem;
 }
 
 export interface TaskFormulaDefinitionDto {
@@ -38,11 +47,17 @@ export interface WorkspaceItemDto {
   /** Date-time string (UTC) */
   deleteAfter?: string;
   members: WorkspaceMemberItem[];
+  projectMemberships: WorkspaceProjectMembershipItem[];
   taskPropertyDefinitions: TaskPropertyDefinitionDto[];
   taskFormulaDefinitions: TaskFormulaDefinitionDto[];
   topRailConfigs: WorkspaceTaskCardLabelConfig[];
   bottomRailConfigs: WorkspaceTaskCardLabelConfig[];
   systemTokens: SystemTokenAccessDto[];
+}
+
+export interface WorkspaceProjectMembershipItem {
+  projectId: string;
+  projectName: string;
 }
 
 export interface WorkspaceTaskCardLabelConfig {